+++ store/streaming/README.md
